+++ packages/ui/Image/Image.js
@@ -8,6 +8,8 @@
 import kind from '@enact/core/kind';
 import PropTypes from 'prop-types';
 import React from 'react';
+import warning from 'warning';
+
 import {selectSrc} from '../resolution';
 
 import componentCss from './Image.less';
@@ -44,16 +46,6 @@ const ImageBase = kind({
 	name: 'ui:Image',
 
 	propTypes: /** @lends ui/Image.Image.prototype */ {
-		/**
-		 * String value or Object of values used to determine which image will appear on
-		 * a specific screenSize.
-		 *
-		 * @type {String|Object}
-		 * @required
-		 * @public
-		 */
-		src: PropTypes.oneOfType([PropTypes.string, PropTypes.object]).isRequired,
-
 		/**
 		 * Sets the aria-label for the image. If unset, it defaults to the value of `alt`
 		 *
@@ -132,12 +124,23 @@ const ImageBase = kind({
 		 * @default 'fill'
 		 * @public
 		 */
-		sizing: PropTypes.oneOf(['fit', 'fill', 'none'])
+		sizing: PropTypes.oneOf(['fit', 'fill', 'none']),
+
+		/**
+		 * String value or Object of values used to determine which image will appear on
+		 * a specific screenSize.
+		 *
+		 * @type {String|Object}
+		 * @required
+		 * @public
+		 */
+		src: PropTypes.oneOfType([PropTypes.string, PropTypes.object])
 	},
 
 	defaultProps: {
 		placeholder: '',
-		sizing: 'fill'
+		sizing: 'fill',
+		src: ''
 	},
 
 	styles: {
@@ -147,10 +150,11 @@ const ImageBase = kind({
 	},
 
 	computed: {
-		bgImage: ({src, placeholder}) => {
-			const imageSrc = selectSrc(src);
+		bgImage: ({placeholder, src}) => {
+			const imageSrc = selectSrc(src) || placeholder;
+			warning(imageSrc, 'Image requires that either the "placeholder" or "src" props be specified.');
 			if (!imageSrc) return null;
-			return placeholder ? `url("${imageSrc}"), url("${placeholder}")` : `url("${imageSrc}")`;
+			return placeholder && imageSrc !== placeholder ? `url("${imageSrc}"), url("${placeholder}")` : `url("${imageSrc}")`;
 		},
 		className: ({className, sizing, styler}) => {
 			return sizing !== 'none' ? styler.append(sizing) : className;