+++ packages/moonstone/Button/Button.js
@@ -42,6 +42,7 @@ const deprecation = deprecate(() => {}, {
  *
  * @class ButtonBase
  * @memberof moonstone/Button
+ * @extends ui/Button.ButtonBase
  * @ui
  * @public
  */
@@ -87,12 +88,7 @@ const ButtonBase = kind({
 		 * @type {Object}
 		 * @public
 		 */
-		css: PropTypes.object,
-
-		/*
-		 * Passed on to UI, referenced here for performance.
-		 */
-		icon: PropTypes.any
+		css: PropTypes.object
 	},
 
 	styles: {
@@ -119,7 +115,7 @@ const ButtonBase = kind({
 			'data-webos-voice-intent': 'Select',
 			...rest,
 			css,
-			iconComponent: (rest.icon ? Icon : null)  // Conditionally send the Icon component to UiButtonBase
+			iconComponent: Icon
 		});
 	}
 });
@@ -143,12 +139,11 @@ const ButtonBase = kind({
  *
  * @hoc
  * @memberof moonstone/Button
- * @extends moonstone/Button.ButtonBase
  * @mixes i18n/Uppercase.Uppercase
  * @mixes moonstone/Marquee.MarqueeDecorator
  * @mixes ui/Button.ButtonDecorator
  * @mixes spotlight/Spottable.Spottable
- * @mixes ui/Skinnable.Skinnable
+ * @mixes moonstone/Skinnable.Skinnable
  * @public
  */
 const ButtonDecorator = compose(