+++ packages/moonstone/Icon/Icon.js
@@ -27,6 +27,7 @@ import componentCss from './Icon.less';
  *
  * @class IconBase
  * @memberof moonstone/Icon
+ * @extends ui/Icon.Icon
  * @ui
  * @public
  */
@@ -136,7 +137,7 @@ const IconBase = kind({
  *
  * @hoc
  * @memberof moonstone/Icon
- * @mixes ui/Skinnable.Skinnable
+ * @mixes moonstone/Skinnable.Skinnable
  * @public
  */
 const IconDecorator = compose(