+++ packages/spotlight/src/spotlight.js
@@ -18,6 +18,7 @@
 
 import {is} from '@enact/core/keymap';
 import {isWindowReady} from '@enact/core/snapshot';
+import platform from '@enact/core/platform';
 import last from 'ramda/src/last';
 
 import Accelerator from '../Accelerator';
@@ -354,6 +355,12 @@ const Spotlight = (function () {
 		_pointerMoveDuringKeyPress = false;
 	}
 
+	function handleWebOSMouseEvent (ev) {
+		if (ev && ev.detail && ev.detail.type === 'Leave') {
+			onBlur();
+		}
+	}
+
 	function onFocus () {
 		// Normally, there isn't focus here unless the window has been blurred above. On webOS, the
 		// platform may focus the window after the app has already focused a component so we prevent
@@ -482,6 +489,7 @@ const Spotlight = (function () {
 				window.addEventListener('keyup', onKeyUp);
 				window.addEventListener('mouseover', onMouseOver);
 				window.addEventListener('mousemove', onMouseMove);
+				if (platform.webos) document.addEventListener('webOSMouse', handleWebOSMouseEvent);
 				setLastContainer(rootContainerId);
 				configureDefaults(containerDefaults);
 				configureContainer(rootContainerId);
@@ -504,6 +512,7 @@ const Spotlight = (function () {
 			window.removeEventListener('keyup', onKeyUp);
 			window.removeEventListener('mouseover', onMouseOver);
 			window.removeEventListener('mousemove', onMouseMove);
+			if (platform.webos) document.removeEventListener('webOSMouse', handleWebOSMouseEvent);
 			Spotlight.clear();
 			_initialized = false;
 		},